Required Changes for Pipeline Integration
