+++ .idea/compiler.xml
@@ -6,8 +6,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="maven-arx" />
+        <module name="maven_arx" />
       </profile>
     </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="arx_attempt" target="16" />
+    </bytecodeTargetLevel>
   </component>
 </project>
