--- maven-arx/pom.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>arx-attempt</artifactId>
-        <groupId>org.example</groupId>
-        <version>1.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>maven-arx</artifactId>
-
-    <properties>
-        <maven.compiler.source>16</maven.compiler.source>
-        <maven.compiler.target>16</maven.compiler.target>
-    </properties>
-
-    <dependencies>
-        <!-- Include the local ARX JAR -->
-        <dependency>
-            <groupId>org.deidentifier</groupId>
-            <artifactId>arx</artifactId>
-            <version>3.9.1</version>
-            <scope>system</scope>
-            <systemPath>/home/kailash/Desktop/arx-anonymization/lib/libarx-3.9.1.jar</systemPath>
-        </dependency>
-    </dependencies>
-
-</project>
