--- maven-arx/src/main/java/call_java.py
@@ -1,8 +1,0 @@
-# call_arx_testing.py
-from java.lang import System
-import ARXTesting
-
-def main():
-    ARXTesting.main([])
-if __name__ == "__main__    ":
-    main()
