--- maven-arx/src/main/java/config.properties
@@ -1,13 +1,0 @@
-dataset.path=/home/kailash/Desktop/arx-anonymization/maven-arx/src/main/java/categoricalDatasetProcessed_fin_TEL_bounds_corrected_1_22.csv
-dataset.charset=UTF-8
-delimiter=,
-
-identifying.columns=Record,Farmer Name,Land Area,Khasra No./ Dag No.,Survey No.
-insensitive.columns=Sample No.,village,pH,EC,OC,N,P,K,S,Zn,Fe,Cu,Mn,B,state,district,subdistrict,cycle
-
-# Configuration for hierarchy levels and interval widths for each column
-# Format: columnName1:numLevels:intervalWidth,columnName2:numLevels:intervalWidth
-columns.hierarchy=Longitude:3:0.5,Latitude:3:0.5
-
-k.anonymity=3
-suppression.limit=0
