+++ maven_arx/src/main/java/config.properties
@@ -2,15 +2,13 @@ dataset.path=/home/kailash/Desktop/arx_anonymization/maven_arx/src/main/java/cat
 dataset.charset=UTF-8
 delimiter=,
 
-insensitive.columns=Sample No.,Farmer Name,village,pH,EC,OC,N,P,K,S,Zn,Fe,Cu,Mn,B,state,district,subdistrict,cycle
-# Suppression and pseudonymization settings
-suppress=Record,Land Area,Khasra No./ Dag No.,Survey No.
-pseudonymize=Farmer Name
-
+identifying.columns=
+insensitive.columns=Record,SampleNo.,FarmerName,village,pH,EC,OC,N,P,K,S,Zn,Fe,Cu,Mn,B,state,district,subdistrict,cycle
+suppress=LandArea,KhasraNo./DagNo.,SurveyNo
+pseudonymize=FarmerName
 # Configuration for hierarchy levels and interval widths for each column
-# Format: columnName1:numLevels:intervalWidth:groupsizes,columnName2:numLevels:intervalWidth:groupsizes
+# Format: columnName1:numLevels:intervalWidth,columnName2:numLevels:intervalWidth
 columns.hierarchy=Longitude:3:0.5:2:2,Latitude:3:0.5:2:2
 
 k.anonymity=3
 suppression.limit=0
-
