+++ pom.xml
@@ -9,7 +9,7 @@
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>maven-arx</module>
+        <module>maven_arx</module>
     </modules>
 
     <properties>
